feat: moved components to the different directory
