+++ src/components/smartComponents/filmCard/FilmCard.jsx
